requirements.txt
setuptools~=65.5.1
Django~=4.1.5
djangorestframework~=3.14.0
sendgrid~=6.9.7
python-dotenv~=0.21.1